# src/main/webapp/doc/swagger.json
{"swagger":"2.0","info":{"description":"Rede Social","version":"1.0.0","title":"Rede Social API"},"host":"localhost:8080","basePath":"/treinamento/api","tags":[{"name":"Controle de Pessoa"}],"schemes":["http"],"paths":{"/pessoas":{"get":{"tags":["Controle de Pessoa"],"summary":"Lista a(s) Pessoa ","description":"","operationId":"obterListaPessoa_1","consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"default":{"description":"successful operation"}}},"post":{"tags":["Controle de Pessoa"],"summary":"Cadastra uma Pessoa ","description":"","operationId":"obterListaPessoa","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/Pessoa"}}],"responses":{"default":{"description":"successful operation"}}},"put":{"tags":["Controle de Pessoa"],"operationId":"atualizarPessoa","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/Pessoa"}}],"responses":{"default":{"description":"successful operation"}}}},"/pessoas/{id}":{"get":{"tags":["Controle de Pessoa"],"operationId":"obterPessoa","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"default":{"description":"successful operation"}}},"delete":{"tags":["Controle de Pessoa"],"operationId":"apagarPessoa","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"default":{"description":"successful operation"}}}}},"definitions":{"Pessoa":{"type":"object","required":["dataNascimento","email","nome","situacao"],"properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"email":{"type":"string"},"dataNascimento":{"type":"string","format":"date"},"situacao":{"type":"boolean"}}}}}